tests: move lib tests to __tests__ dir.
