+++ src/__tests__/unit/lib/case-studies/emem.test.ts
@@ -1,5 +1,5 @@
 import {describe, expect, jest, test} from '@jest/globals';
-import {EMemModel} from './emem-model';
+import {EMemModel} from '../../../../lib/case-studies/emem-model';
 
 jest.setTimeout(30000);
