+++ src/__tests__/unit/lib/ccf/index.test.ts
@@ -1,5 +1,5 @@
 import {describe, expect, jest, test} from '@jest/globals';
-import {CloudCarbonFootprint} from './index';
+import {CloudCarbonFootprint} from '../../../../lib/ccf/index';
 
 jest.setTimeout(30000);
