+++ src/__tests__/unit/lib/sci-o/index.test.ts
@@ -1,5 +1,5 @@
 import {describe, expect, jest, test} from '@jest/globals';
-import {SciOModel} from './index';
+import {SciOModel} from '../../../../lib/sci-o/index';
 jest.setTimeout(30000);
 
 describe('ccf:configure test', () => {
