+++ src/__tests__/unit/lib/sci/index.test.ts
@@ -1,5 +1,5 @@
 import {describe, expect, jest, test} from '@jest/globals';
-import {SciModel} from './index';
+import {SciModel} from '../../../../lib/sci/index';
 jest.setTimeout(30000);
 
 describe('sci:configure test', () => {
