+++ src/__tests__/unit/lib/watt-time/index.test.ts
@@ -1,6 +1,6 @@
 import {describe, expect, jest, test} from '@jest/globals';
-import {WattTimeGridEmissions} from './index';
-import * as DATA from '../../__mocks__/watt-time/data.json';
+import {WattTimeGridEmissions} from '../../../../lib/watt-time/index';
+import * as DATA from '../../../../__mocks__/watt-time/data.json';
 import axios from 'axios';
 
 jest.setTimeout(30000);
